Try configuring non-blocking on incoming/outgoing connectors before connecting
We were unconditionally setting blocking=false (non-blocking) just after binding
or connecting to the socket.

This was the source of some strange and hard to reproduce ioctl failures on
Windows. My suspicion is that we are opening a socket, connecting with it and
then before we set non-blocking, something is using the socket in a way
that prevents us from setting non-blocking.

This change moves the non-blocking setup to before the connect/bind, so there is
no opportunity for anything else to use the socket.

refs #30208