Remove cycle between ResolveContext and DefaultTransformUpstreamDependenciesResolver
Previously, DefaultTransformUpstreamDependenciesResolver relied on the strictResovlerResults of a ResolveContext. These
strictResovlerResults were actually a memoized object _containing_ the same DefaultTransformUpstreamDependenciesResolver.
In effect, the resolver was constructed with a future to itself. This cycle can make it very difficult to implement
and initiate resolution from anything other than a Configuration, the sole implementation of ResolveContext.

The strict resolver results provides two values. The partial value and the complete value. We also construct a transform dependencies
resolver twice for a given configuration, once when computing the partial value and once when computing the complete value.

When 'pulling on' the dependencies resolver for the partial result, we either look at the partial results from the graph, or need the
complete results. When 'pulling' on the dependenciesr resolver for the complete set, we need only the complete results from the graph.

For this reason, we can implement the complete results independently. It needs no prior work to be done, and it has no other dependenices
other than the normal inputs of resolution. The partial results, however, must receive a _future_ to the complete results. Note, it no longer
receives a future to itself. The code has been reworked to pass in the partial results eagerly, as they are already known at the time of
constructing the transform dependencies resolver.

While we have gotten rid of a major logic/object cycle, there still remains a cycle. This loop is much 'smaller' but remains nonetheless.
A DefaultVisistedArtifactSet requires a TransformUpstreamDependenciesResolver and a DefaultTransformUpstreamDependenciesResolver requires a
VisitedArtifactSet. This cycle existed before, but as part of the larger cycle. We have simply reduced the size of the loop.

Some additional work is needed to simplify DefaultTransformUpstreamDependenciesResolver, as the logic is very complicated, and this commit
did the minimum possible changes to remove the cycle. These minimum changes are a bit ugly.