Support creating root variant state from configuration instance
We currently support the edge case where the user may remove the configuration from the buildscript
classpath configuration from the configuration container. However, the script handler keeps a reference
to that configuration and still tries to resolve it. We previously then tried to lookup that configuration
in the container by name, which threw an exception.

Removing a configuration from the buildscript configuration container is depreceated, so we can revert all of
this in 9.0